+++ src/components/navleft.vue
@@ -44,10 +44,10 @@
             <el-submenu index="4">
               <template slot="title"><span>vuex数据仓库</span></template>
               <el-menu-item-group>
-                <el-menu-item index="countpage">count数字加减</el-menu-item>
+                <el-menu-item index="vuex1">vuex1数字加减</el-menu-item>
+                <el-menu-item index="vuex3">vuex3</el-menu-item>
+                <el-menu-item index="vuex4">vuex4输出状态</el-menu-item>
                 <el-menu-item index="countpage2">锚点跳转</el-menu-item>
-                <el-menu-item index="countpage3">vuex3</el-menu-item>
-                <el-menu-item index="countpage4">vuex4</el-menu-item>
               </el-menu-item-group>
             </el-submenu>
           </el-menu>
