+++ src/pages/vuex1.vue
@@ -1,11 +1,9 @@
 <template>
   <div>
-    <h1>{{ msg }}</h1><hr>
     <h3>{{count}}</h3>
     <p>
       <button @click="$store.commit('add',10)">+</button>
       <button @click="$store.commit('reduce')">-</button>
-      <button @click="aaa">aaa</button>
       
     </p>
   </div>
@@ -16,8 +14,7 @@
   export default {
     data () {
       return {
-        msg:'hello vuex',
-        arr:'',
+        msg:'hello vuex',  
       };
     },
     computed:{
@@ -26,16 +23,5 @@
       }
     },
     store,
-    methods:{
-      aaa(){
-        let json = {
-          a1:123,
-          a2:'张三',
-          a3:'李四'
-        }
-        this.arr =json
-        console.log(arr)
-      }
-    }
   };
 </script>
